Fix compile warnnings
merge fix into main

Fix compile warnnings

Created-by: Ooohui
Commit-by: Ooohui
Merged-by: liwei3013
Description: Fix compile warnnings

See merge request: openHiTLS/openhitls!151


# message auto-generated for no-merge-commit merge:
merge openhitls-0.1 into openhitls-0.1

Fix compile warnnings

Created-by: Ooohui
Commit-by: Ooohui
Merged-by: liwei3013
Description: Fix compile warnnings

See merge request: openHiTLS/openhitls!154

+++ bsl/asn1/src/bsl_asn1_print.c
@@ -71,7 +71,7 @@ int32_t BSL_ASN1_Printf(uint32_t layer, BSL_UIO *uio, const char *fmt, ...)
         BSL_ERR_PUSH_ERROR(BSL_ASN1_ERR_PRINTF);
         return BSL_ASN1_ERR_PRINTF;
     }
-    int32_t ret = BSL_ASN1_PrintfBuff(layer, uio, buff, strlen(buff));
+    int32_t ret = BSL_ASN1_PrintfBuff(layer, uio, buff, (uint32_t)strlen(buff));
     va_end(args);
 
     return ret;
